File renaming and restructure for ChatAPI to make everything more readable
