+++ ChatAPI/src/db/searchEmbed.ts
@@ -1,15 +1,7 @@
 import { OpenAIEmbeddings } from "@langchain/openai";
-import pkg from 'pg';
 import 'dotenv/config'
 
-const { Pool } = pkg;
-const POOL = new Pool({
-    user: process.env.DB_USER,
-    password: process.env.DB_PWD,
-    host: process.env.DB_HOST,
-    database: process.env.DB_DATABASE_NAME,
-    port: process.env.DB_PORT
-});
+import { pool } from "./pool.ts"
 
 const OPENAI_CALL_EMBED = new OpenAIEmbeddings({
     openAIApiKey: process.env.OPENAI_API_KEY,
@@ -53,7 +45,7 @@ async function createEmbedQuery(text: string): Promise<OpenAIEmbedResult> {
 
 export async function searchAudioEmbed({ videoID, query }: SearchDBParms): Promise<{ videoID: string, Audios: DBEmbedResult[] }> {
     let embedQuery = await createEmbedQuery(query);
-    let client = await POOL.connect();
+    let client = await pool.connect();
     let sqlQueryAudio = `SELECT clipId, rawText, startTime, endTime FROM audio_embeds WHERE videoId = '${videoID}' ORDER BY embedding <-> '[${embedQuery.queryEmbed.toString()}]' LIMIT 15`
     let resAudio = await client.query(sqlQueryAudio);
     client.release();
@@ -70,7 +62,7 @@ export async function searchAudioEmbed({ videoID, query }: SearchDBParms): Promi
 
 export async function searchVisualEmbed({ videoID, query }: SearchDBParms): Promise<{ videoID: string, Frames: DBEmbedResult[] }> {
     let embedQuery = await createEmbedQuery(query);
-    let client = await POOL.connect();
+    let client = await pool.connect();
     let sqlQueryFrames = `SELECT clipId, rawText, startTime, endTime, frameId FROM frame_embeds WHERE videoId = '${videoID}' ORDER BY embedding <-> '[${embedQuery.queryEmbed.toString()}]'  LIMIT 15`
     let resFrames = await client.query(sqlQueryFrames);
     client.release();
@@ -87,7 +79,7 @@ export async function searchVisualEmbed({ videoID, query }: SearchDBParms): Prom
 
 export async function searchVisualEmbedForImages({ videoID, query }: SearchDBParms): Promise<{ videoID: string, Frames: DBEmbedImageResult[] }> {
     let embedQuery = await createEmbedQuery(query);
-    let client = await POOL.connect();
+    let client = await pool.connect();
     let sqlQueryFrames = `SELECT frame_embeds.clipId, imgurl, startTime, endTime, frame_embeds.frameId FROM frame_embeds, s3_files_frame WHERE frame_embeds.videoId = '${videoID}' AND s3_files_frame.clipId = frame_embeds.clipId AND s3_files_frame.frameId = frame_embeds.frameId  ORDER BY embedding <-> '[${embedQuery.queryEmbed.toString()}]' LIMIT 15`
     let resFrames = await client.query(sqlQueryFrames);
     client.release();
@@ -101,7 +93,6 @@ export async function searchVisualEmbedForImages({ videoID, query }: SearchDBPar
         Frames: resFrames.rows,
     };
 }
-
 
 export function convertDBEmbedResultToString(embeds: DBEmbedResult[]): string {
     let textResult = "";
